# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "教务系统",
        "description": "教务系统接口文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/teacher/student_list": {
            "get": {
                "tags": [
                    "课程管理"
                ],
                "summary": "学生列表",
                "operationId": "55455eda2a105269c0c86fad7ce913d1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "meta": {
                                                    "description": "元信息",
                                                    "properties": {
                                                        "count": {
                                                            "description": "当前页的项目数",
                                                            "type": "integer"
                                                        },
                                                        "perPage": {
                                                            "description": "每页显示的项目数",
                                                            "type": "integer"
                                                        },
                                                        "currentPage": {
                                                            "description": "当前页码",
                                                            "type": "integer"
                                                        },
                                                        "lastPage": {
                                                            "description": "最后一页的页码",
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "description": "总数",
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "list": {
                                                    "description": "数据列表",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "唯一标识",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "姓名",
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "description": "邮箱",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teacher/get_not_course_student_list": {
            "get": {
                "tags": [
                    "课程管理"
                ],
                "summary": "获取未选修指定科目的学生列表",
                "operationId": "233ad6e75ca7601d7ec5a99846656d21",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "query",
                        "description": "课程id"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "meta": {
                                                    "description": "元信息",
                                                    "properties": {
                                                        "count": {
                                                            "description": "当前页的项目数",
                                                            "type": "integer"
                                                        },
                                                        "perPage": {
                                                            "description": "每页显示的项目数",
                                                            "type": "integer"
                                                        },
                                                        "currentPage": {
                                                            "description": "当前页码",
                                                            "type": "integer"
                                                        },
                                                        "lastPage": {
                                                            "description": "最后一页的页码",
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "description": "总数",
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "list": {
                                                    "description": "数据列表",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "唯一标识",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "姓名",
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "description": "邮箱",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teacher/get_course_student_list": {
            "get": {
                "tags": [
                    "课程管理"
                ],
                "summary": "获取选修指定科目的学生列表",
                "operationId": "2ff3d346a03ebd7599546544d3299a45",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "query",
                        "description": "课程id"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "meta": {
                                                    "description": "元信息",
                                                    "properties": {
                                                        "count": {
                                                            "description": "当前页的项目数",
                                                            "type": "integer"
                                                        },
                                                        "perPage": {
                                                            "description": "每页显示的项目数",
                                                            "type": "integer"
                                                        },
                                                        "currentPage": {
                                                            "description": "当前页码",
                                                            "type": "integer"
                                                        },
                                                        "lastPage": {
                                                            "description": "最后一页的页码",
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "description": "总数",
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "list": {
                                                    "description": "数据列表",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "唯一标识",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "姓名",
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "description": "邮箱",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization-Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}